connecting to database - issue with database credentials
