+++ api/go.mod
@@ -2,7 +2,10 @@ module api
 
 go 1.17
 
-require github.com/stretchr/testify v1.7.0
+require (
+	github.com/lib/pq v1.10.3
+	github.com/stretchr/testify v1.7.0
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.0 // indirect
