--- postgres/go.mod
@@ -1,5 +1,0 @@
-module postgres
-
-go 1.17
-
-require github.com/lib/pq v1.10.3
